Separete content from function
